src/requirements.txt
BingTranslator==0.1
aiohttp==3.5.4
aioredis==1.1.0
asyncpg==0.16.0
bs4==0.0.1
discord.py==1.0.0
feedparser==5.2.1
lxml==4.2.1
parsedatetime==2.4
pytz==2018.4
websockets==6.0
wolframalpha==3.0.1
requests==2.17.3
lxml==4.2.1
pandas==0.23.4
cfscrape==1.9.5
